[librbd] Add hot area cache
Implement a new cache plugin. I call it as hot area cache (Abbr: HAC).

This is a persistent cache technology. It can decrease the VM's read
operation's delay and decrease the traffic pressure of the ceph public
network.

It as a librbd's plugin be implemented. In the plugin, by these steps as
below to determine which datas should be cache.

1. Split the whole image as multiple areas which have same size. When the
user's read extent hit one or more areas, these areas's access count will
be added one.

2. Create a file and split it as multiple slots which of size equal to area's
in step 1.

3. Sort these areas by their count attribute from big to small.